requirements.txt
aiohttp
telethon
cryptg
python-dotenv
pymongo
